pyproject.toml
[tool.poetry]
name = "elecciones-test"
version = "0.1.0"
description = "Test para entrevista Devecoop 2023"
authors = ["FedeG <federico.gonzalez@devecoop.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "elecciones_test"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.7"
django-extensions = "^3.2.3"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-configurations = "^2.5"


[tool.poetry.group.dev.dependencies]
flake8 = "3.8.4"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
ipython = "^8.17.2"
ipdb = "^0.13.13"
pytest-django = "^4.7.0"
flake8-django = "^1.4"
factory-boy = "^3.3.0"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.blue]
line-length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'ballot.settings'
addopts = "-s --no-migrations --disable-warnings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
